Hasta B6

+++ bloque5/bloque5.py
@@ -15,7 +15,7 @@ STORE_DIR = "rag_store"
 DESC_FILE = "bloque4/description.json"
 TOP_K = 2
 LOG_FILE = "results/agent_log_b5.txt"
-os.remove(LOG_FILE)
+if os.path.exists(LOG_FILE): os.remove(LOG_FILE)
 with open(LOG_FILE, "w", encoding="utf-8") as f:
         f.write("LOGS:\n")
 # =============================
